Btn return in list cells for each sectors and move files
